refactor main variants
- move log_end to uitl.rs
- move CliError to its own module - clieerror.rs
- create additional CliError enum - NoMatch()
