+++ src/clierror.rs
@@ -1,0 +1,66 @@
+use std::borrow::ToOwned;
+use std::fmt;
+use std::io;
+
+#[derive(Debug)]
+pub enum CliError {
+    Flag(docopt::Error),
+    Csv(csv::Error),
+    Io(io::Error),
+    NoMatch(),
+    Other(String),
+}
+
+impl fmt::Display for CliError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match *self {
+            CliError::Flag(ref e) => e.fmt(f),
+            CliError::Csv(ref e) => e.fmt(f),
+            CliError::Io(ref e) => e.fmt(f),
+            CliError::NoMatch() => f.write_str("no_match"),
+            CliError::Other(ref s) => f.write_str(s),
+        }
+    }
+}
+
+impl From<docopt::Error> for CliError {
+    fn from(err: docopt::Error) -> CliError {
+        CliError::Flag(err)
+    }
+}
+
+impl From<csv::Error> for CliError {
+    fn from(err: csv::Error) -> CliError {
+        if !err.is_io_error() {
+            return CliError::Csv(err);
+        }
+        match err.into_kind() {
+            csv::ErrorKind::Io(v) => From::from(v),
+            _ => unreachable!(),
+        }
+    }
+}
+
+impl From<io::Error> for CliError {
+    fn from(err: io::Error) -> CliError {
+        CliError::Io(err)
+    }
+}
+
+impl From<String> for CliError {
+    fn from(err: String) -> CliError {
+        CliError::Other(err)
+    }
+}
+
+impl<'a> From<&'a str> for CliError {
+    fn from(err: &'a str) -> CliError {
+        CliError::Other(err.to_owned())
+    }
+}
+
+impl From<regex::Error> for CliError {
+    fn from(err: regex::Error) -> CliError {
+        CliError::Other(format!("Regex error: {err:?}"))
+    }
+}
